Create housekeeping2.py
Added the calc of the means and std in a function so we can redo them easily when we add image augmentation in future. Also my results for mean and std is slightly different here, could be cause i calc mean and std after resizing the image? 


Check:

this is avg RGB of the first image in dataset: Abyssinian_1 using this website (https://10015.io/tools/image-average-color-finder)

(74, 76, 59) / 255 =
[0.290, 0.298, 0.231]

result with this code:

[0.2911, 0.2981, 0.2294]